+++ DecisionTreeRegressor/split_functions.py
@@ -1,0 +1,50 @@
+from cost_functions import mean_squared_error
+import pandas as pd
+
+### --- Split Functions --- ###
+
+def best_split_numerical(self, X:pd.DataFrame, y:pd.Series) -> tuple:
+    """
+    Find the best split for a group of numerical columns. Returning
+    the feature to split on along with the threshold to take
+
+    Args:
+    X (pd.DataFrame): Data frame of features and their values
+    y (pd.Series): Target variable for the same dataset
+
+    Returns:
+        tuple: (feature, threshold)
+    """
+
+    # All features
+    features = X.columns
+    n_samples = X.shape[0]
+
+    # Current best splits and thresholds
+    best_feature, best_threshold, best_mse = None, None, float("inf")
+
+    # Iterate through each feature and potential threshold
+    for f in features:
+        unique_f_values = X[:,f].unique()
+        # Take all potential values as thresholds for now
+        for threshold in unique_f_values:
+            # Find the values either side of the split
+            left_mask = X[:,f] <= threshold
+            right_mask = ~left_mask
+
+            # Avoid empty splits on either side by checking true counts
+            if sum(left_mask) == 0 or sum(right_mask) == 0:
+                continue
+
+            # Calculate the MSE on either side of the split
+            # The lower the variance the higher the homogeneity 
+            left_mse = mean_squared_error(y[left_mask])
+            right_mse = mean_squared_error(y[right_mask])
+            # Weighted mse dependent on split sizes
+            total_mse = ((sum(left_mask)*left_mse) + (sum(right_mask)*right_mse))/n_samples
+
+            # Check if this MSE outperforms current
+            if total_mse < best_mse:
+                best_feature, best_threshold, best_mse = f, threshold, total_mse
+    
+    return best_feature, best_threshold
